remake App location
