+++ src/components/App/App.module.scss
@@ -10,12 +10,9 @@
   width: 100%;
 }
 
-.burger-container {
+.burgerContainer {
   display: flex;
   justify-content: space-between;
-}
-.content-wrapper {
-  max-width: 1240px;
 }
 
 .text {
