+++ src/index.css
@@ -23,3 +23,7 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+
+.content-wrapper {
+  max-width: 1240px;
+}
